Move bench compare tests beside implementation
